# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MAIDERES API
  version: 1.0.0
  description: |
    API cœur de la marketplace MAIDERES (Cameroun, Douala). Consommée par
    `apps/web` (console back-office) et `maidere-connect` (vitrine + espaces
    client/prestataire).

    ## Versioning
    Toute route est montée deux fois : sans préfixe de version (`/api/...`,
    conservé pour compatibilité avec les consommateurs existants) et sous
    `/api/v1/...` (préfixe recommandé pour toute nouvelle intégration). Les
    deux pointent vers exactement le même code — aucune divergence de
    comportement aujourd'hui. Un futur `v2` ne casserait que les routes
    explicitement listées comme changées, jamais silencieusement.

    ## Auth
    Bearer JWT Supabase (`Authorization: Bearer <access_token>`), sauf le
    groupe `Public` (aucun token). Le rôle technique (`profiles.role`) ne
    suffit pas à lui seul : l'identité métier (staff / prestataire / client)
    est dérivée de l'existence d'une ligne `prestataires`/`clients` liée au
    profil — voir `GET /profile/me`.

    ## Enveloppe de réponse
    Succès : `{ "data": ... }` (collections : `{ "data": [...] }`, parfois
    `meta.page/pageSize/total`). Erreur : `{ "error": "...", "code"?: "..." }`.
  contact:
    name: MAIDERES
...
servers:
  - url: /api/v1
    description: Préfixe versionné (recommandé)
  - url: /api
    description: Alias non versionné (compatibilité)
security:
  - bearerAuth: []
...
paths:
...
  /public/prestataires:
    get:
      tags: [Public]
      summary: Annuaire public des prestataires actifs
      security: []
      parameters:
        - { name: ville, in: query, schema: { type: string } }
        - { name: quartier, in: query, schema: { type: string } }
        - { name: categorie, in: query, schema: { type: string }, description: Filtre sur `metier`. }
        - { name: recherche, in: query, schema: { type: string }, description: Recherche partielle sur `nom` (ILIKE). }
      responses:
        '200':
          description: Jusqu'à 60 résultats, triés note_moyenne desc puis nom.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { type: array, items: { $ref: '#/components/schemas/PrestatairePublic' } } }
...
  /public/promotions:
    get:
      tags: [Public]
      summary: Promotions actives, tous prestataires actifs confondus
      security: []
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 60 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { data: { type: array, items: { $ref: '#/components/schemas/Promotion' } } }

  # ── Profile ──────────────────────────────────────────────────────────────
...
  /categories_services:
    get:
      tags: [Categories]
      summary: Lister — staff&#58; toutes ; sinon&#58; actif=true seulement
      responses: { '200': { $ref: '#/components/responses/Many' } }
    post:
      tags: [Categories]
      summary: Créer (admin uniquement)
      requestBody: { $ref: '#/components/requestBodies/CreateCategorieService' }
      responses: { '201': { $ref: '#/components/responses/One' } }
...
  /prestataires:
    get:
      tags: [Prestataires]
      summary: Lister — client&#58; statut=actif seulement
      responses: { '200': { $ref: '#/components/responses/Many' } }
    post:
      tags: [Prestataires]
      summary: Auto-inscription (statut forcé à en_attente côté serveur)
      requestBody: { $ref: '#/components/requestBodies/CreatePrestataire' }
      responses: { '201': { $ref: '#/components/responses/One' } }
...
  /clients:
    get:
      tags: [Clients]
      responses: { '200': { $ref: '#/components/responses/Many' } }
    post:
      tags: [Clients]
      summary: Auto-inscription (ou création par staff via profile_id)
      requestBody: { $ref: '#/components/requestBodies/CreateClient' }
      responses: { '201': { $ref: '#/components/responses/One' } }
...
  /demandes:
    get:
      tags: [Demandes]
      summary: Lister — client&#58; les siennes ; prestataire&#58; celles matchées avec lui ; staff&#58; toutes
      responses: { '200': { $ref: '#/components/responses/Many' } }
    post:
      tags: [Demandes]
      summary: Créer (self-service ou staff pour un client donné, multi-canal)
      requestBody: { $ref: '#/components/requestBodies/CreateDemande' }
      responses: { '201': { $ref: '#/components/responses/One' } }
...
  /matchings:
    get:
      tags: [Matchings]
      responses: { '200': { $ref: '#/components/responses/Many' } }
    post:
      tags: [Matchings]
      summary: Proposer un prestataire pour une demande (staff uniquement)
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ProposeMatchingInput' } }
      responses: { '201': { $ref: '#/components/responses/One' } }
...
  /avis:
    get:
      tags: [Avis]
      responses: { '200': { $ref: '#/components/responses/Many' } }
    post:
      tags: [Avis]
      summary: Le client laisse un avis sur son propre matching 'realise'
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/CreateAvisInput' } }
      responses: { '201': { $ref: '#/components/responses/One' } }
...
  /transactions:
    get:
      tags: [Transactions]
      responses: { '200': { $ref: '#/components/responses/Many' } }
    post:
      tags: [Transactions]
      summary: Staff, uniquement pour un matching 'realise'. commission_taux/montant toujours calculés par trigger.
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/CreateTransactionInput' } }
      responses: { '201': { $ref: '#/components/responses/One' } }
...
  /commission_config:
    get:
      tags: [Commission Config]
      summary: Staff uniquement (lecture) — categorie_id null = règle globale/fallback
      responses: { '200': { $ref: '#/components/responses/Many' } }
    post:
      tags: [Commission Config]
      summary: Admin uniquement
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/CreateCommissionConfigInput' } }
      responses: { '201': { $ref: '#/components/responses/One' } }
...
  /offres:
    get:
      tags: [Offres]
      responses: { '200': { $ref: '#/components/responses/Many' } }
    post:
      tags: [Offres]
      summary: Self-service (publie=true par défaut, non réglable à la création)
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/CreateOffreInput' } }
      responses: { '201': { $ref: '#/components/responses/One' } }
...
  /promotions:
    get:
      tags: [Promotions]
      responses: { '200': { $ref: '#/components/responses/Many' } }
    post:
      tags: [Promotions]
      summary: Self-service. debut = now() par défaut, active toujours true à la création.
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/CreatePromotionInput' } }
      responses: { '201': { $ref: '#/components/responses/One' } }
...
  /realisations:
    get:
      tags: [Realisations]
      responses: { '200': { $ref: '#/components/responses/Many' } }
    post:
      tags: [Realisations]
      summary: Self-service. image_url pointe vers le bucket Storage `maideres`.
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/CreateRealisationInput' } }
      responses: { '201': { $ref: '#/components/responses/One' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase JWT
...
  responses:
    One:
      description: Objet unique
      content:
        application/json:
          schema:
            type: object
            properties: { data: {} }
    Many:
      description: Collection
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { type: array, items: {} }
              meta:
                type: object
                properties: { page: { type: integer }, pageSize: { type: integer }, total: { type: integer } }
...
  requestBodies:
...
    CreateCategorieService:
      required: true
      content:
        application/json: { schema: { $ref: '#/components/schemas/CreateCategorieServiceInput' } }
...
    CreatePrestataire:
      required: true
      content:
        application/json: { schema: { $ref: '#/components/schemas/CreatePrestataireInput' } }
...
    CreateClient:
      required: true
      content:
        application/json: { schema: { $ref: '#/components/schemas/CreateClientInput' } }
...
    CreateDemande:
      required: true
      content:
        application/json: { schema: { $ref: '#/components/schemas/CreateDemandeInput' } }
  schemas:
...
    CreateCategorieServiceInput:
      type: object
      required: [libelle]
      properties:
        libelle: { type: string, minLength: 1, maxLength: 100 }
        actif: { type: boolean, default: true }
...
    PrestatairePublic:
      type: object
      description: Vue exposée à la vitrine — jamais profile_id/geoloc/categories/taux_commission.
      properties:
        id: { type: string, format: uuid }
        nom: { type: string }
        telephone: { type: string }
        quartier: { type: string, nullable: true }
        ville: { type: string, nullable: true }
        metier: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        disponible: { type: boolean }
        zones_couverture: { type: array, items: { type: string } }
        note_moyenne: { type: string }
        statut: { $ref: '#/components/schemas/PrestataireStatut' }
    PrestataireStatut: { type: string, enum: [en_attente, actif, suspendu] }
    CreatePrestataireInput:
      type: object
      required: [nom, telephone]
      properties:
        nom: { type: string, minLength: 1, maxLength: 100 }
        telephone: { type: string, minLength: 6, maxLength: 30 }
        categories: { type: array, items: { type: string, format: uuid }, default: [] }
        quartier: { type: string, nullable: true, maxLength: 100 }
        geoloc_lat: { type: number, minimum: -90, maximum: 90, nullable: true }
        geoloc_lng: { type: number, minimum: -180, maximum: 180, nullable: true }
        ville: { type: string, nullable: true, maxLength: 100 }
        metier: { type: string, nullable: true, maxLength: 100 }
        bio: { type: string, nullable: true, maxLength: 2000 }
        zones_couverture: { type: array, items: { type: string, maxLength: 100 } }
        profile_id: { type: string, format: uuid, description: 'Staff seulement — créer pour un autre profil.' }
...
    CreateClientInput:
      type: object
      required: [nom, telephone]
      description: 'niu requis si type_client = entreprise|organisation (règle applicative).'
      properties:
        nom: { type: string, minLength: 1, maxLength: 100 }
        telephone: { type: string, minLength: 6, maxLength: 30 }
        quartier: { type: string, nullable: true, maxLength: 100 }
        type_client: { type: string, enum: [particulier, entreprise, organisation], default: particulier }
        niu: { type: string, nullable: true, maxLength: 50 }
        whatsapp: { type: string, nullable: true, maxLength: 30 }
        email: { type: string, format: email, nullable: true, maxLength: 150 }
        source: { type: string, enum: [whatsapp, appel, ecommerce, referral], default: whatsapp }
        profile_id: { type: string, format: uuid, description: 'Staff seulement.' }
...
    CreateDemandeInput:
      type: object
      required: [categorie_id, description]
      description: 'Si niveau_urgence=planifie, date_souhaitee devient requise.'
      properties:
        client_id: { type: string, format: uuid, description: 'Staff seulement.' }
        categorie_id: { type: string, format: uuid }
        description: { type: string, minLength: 1, maxLength: 2000 }
        localisation: { type: string, nullable: true, maxLength: 200 }
        canal: { type: string, enum: [web, whatsapp, manuel], default: web }
        niveau_urgence: { type: string, enum: [immediate, urgent, planifie], default: urgent }
        date_souhaitee: { type: string, format: date-time, nullable: true }
...
    ProposeMatchingInput:
      type: object
      required: [demande_id, prestataire_id]
      properties:
        demande_id: { type: string, format: uuid }
        prestataire_id: { type: string, format: uuid }
...
    CreateAvisInput:
      type: object
      required: [matching_id, note]
      properties:
        matching_id: { type: string, format: uuid }
        note: { type: integer, minimum: 1, maximum: 5 }
        commentaire: { type: string, nullable: true, maxLength: 1000 }
...
    CreateTransactionInput:
      type: object
      required: [matching_id, montant_service]
      properties:
        matching_id: { type: string, format: uuid }
        montant_service: { type: integer, minimum: 1 }

    # ── Reversements ──
...
    CreateCommissionConfigInput:
      type: object
      required: [type, valeur]
      description: 'valeur ≤ 100 obligatoire si type=pourcentage.'
      properties:
        categorie_id: { type: string, format: uuid, nullable: true }
        type: { type: string, enum: [pourcentage, montant_fixe] }
        valeur: { type: number, minimum: 0 }
        actif: { type: boolean, default: true }
...
    CreateOffreInput:
      type: object
      required: [categorie, titre, prix]
      properties:
        categorie: { type: string, minLength: 1, maxLength: 100 }
        titre: { type: string, minLength: 1, maxLength: 150 }
        description: { type: string, nullable: true, maxLength: 2000 }
        prestations: { type: array, items: { type: string, maxLength: 150 }, default: [] }
        prix: { type: integer, minimum: 0 }
        unite_prix: { type: string, default: forfait, maxLength: 30 }
        delai_heures: { type: integer, nullable: true, minimum: 1 }
        prestataire_id: { type: string, format: uuid, description: 'Staff seulement.' }
...
    Promotion:
      type: object
      properties:
        id: { type: string, format: uuid }
        prestataire_id: { type: string, format: uuid }
        offre_id: { type: string, format: uuid, nullable: true }
        titre: { type: string }
        description: { type: string, nullable: true }
        remise_pct: { type: string }
        debut: { type: string, nullable: true }
        fin: { type: string, nullable: true }
        active: { type: boolean }
        created_at: { type: string }
    CreatePromotionInput:
      type: object
      required: [titre, remise_pct]
      description: 'debut = now() par défaut, active toujours true à la création.'
      properties:
        offre_id: { type: string, format: uuid, nullable: true }
        titre: { type: string, minLength: 1, maxLength: 150 }
        description: { type: string, nullable: true, maxLength: 2000 }
        remise_pct: { type: number, minimum: 0.01, maximum: 100 }
        fin: { type: string, format: date-time, nullable: true }
        prestataire_id: { type: string, format: uuid, description: 'Staff seulement.' }
...
    CreateRealisationInput:
      type: object
      required: [image_url]
      description: 'image_url pointe vers le bucket Storage `maideres`.'
      properties:
        titre: { type: string, nullable: true, maxLength: 150 }
        description: { type: string, nullable: true, maxLength: 2000 }
        image_url: { type: string, minLength: 1, maxLength: 2000 }
        prestataire_id: { type: string, format: uuid, description: 'Staff seulement.' }

    # ── Notifications (interne, non exposé en écriture) ──
...